Added exception handler to rest-api
