+++ src/main/resources/application.conf
@@ -3,12 +3,12 @@ pipewrench {
     url = "https://github.com/Cargill/pipewrench"
   }
   scripts {
-    install = "conf/requirements.sh"
+    install = "src/main/resources/requirements.sh"
   }
   directory {
-    base = "pipewrench"
-    pipewrench = "pipewrench"
-    templates = "pipewrench/pipewrench/templates"
-    ingest = "pipewrench/ingest_conf"
+    base = "pipewrench_conf"
+    pipewrench = "pipewrench_conf"
+    templates = "pipewrench_conf/pipewrench/templates"
+    ingest = "pipewrench_conf/ingest_conf"
   }
 }
